requirements.txt
streamlit>=1.22.0
numpy>=1.24.3
pandas>=2.0.3
matplotlib>=3.7.2
opencv-python>=4.8.0
torch>=2.0.1
torchvision>=0.15.2
shap>=0.42.1
scikit-learn>=1.3.0
pillow>=10.0.0
lime>=0.2.0.1
scikit-image>=0.20.0
scipy>=1.10.1
ultralytics>=8.0.145
